[mlir] Rework verifyOperation to address scaling issues.
Fixes performance on blocks with many operations where this method
(and not the verification it's driving) is where most CPU time is
spent, managing its worklist datastructures.

The primary culprit is the DenseSet used to track `seen` during the
DFS through the IR structure.  This becomes very expensive on large
IR as the set includes all operations and blocks visited.
Dropping this in favor of a bit on the worklist item is a huge win
addressing the primary scaling issue with this method.

While visiting, move to using a stack of iterators with bit of state
to keep the datastructures modest in size (and cost) even when
verifying large IR containing blocks with many operations.